REMOVE KEYCODE HIGHLIGHTING + CLEANUP
There are just too many keycodes to keep a realistic list of all of
them, especially when you include unicode. This exacerbates the problem
that alias keycodes for common keys ('@', '\\)' etc.) can't be set
properly as keycodes (see previous commits...) and so don't highlight
properly. Maybe there's a way to do it, but it is beyonnd my first
syntax file.

This also makes the file much smaller and simpler to maintain!